requirements.txt
# Core Dependencies
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
web3==6.11.3
python-dotenv==1.0.0

# Blockchain Analysis
eth-utils==2.3.1
eth-abi==4.2.1
hexbytes==0.3.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Visualization
plotly==5.18.0
networkx==3.2.1

# Database
sqlalchemy==2.0.23

# Security
cryptography==41.0.7

# API Rate Limiting
ratelimit==2.2.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
colorama==0.4.6